requirements.txt
admin-totals==1.0.1
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-admin-rangefilter==0.8.1
django-autocomplete-light==3.8.2
django-bootstrap-datepicker-plus==3.0.5
django-crispy-forms==1.12.0
django-filter==2.4.0
django-object-toolbar-admin==0.1.1
django-phonenumber-field==5.2.0
django-smart-selects==1.5.9
django-static-fontawesome==5.14.0.0
djangorestframework==3.12.4
idna==3.2
phonenumbers==8.12.29
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
twilio==6.63.0
urllib3==1.26.6
